DPO3DPKRT-899/Asset File Validation (#658)
(new) endpoint and function for reportAssetFiles
(new) shared enum ProcessState
(new) collecting of assets and asset versions
(new) analysis checking for file size, on disk, etc.
(new) helper type for operation results OpResult
(new) helper enum for ProcessStatus
(new) helper routines for generateGUID
(new) storage exposure routines to get repository path (with version)
(new) split batch generation into own component/file
(new) Admin tools now supports multiple collapsible areas
(new) server-side routines for CSV formatting of asset results
(new) convertFileSizeToString helper added making sizes human readable
(new) DataTableSelect component for selectable data
(new) ToolsAssetValidation skeleton component
(new) API.getReport routine for getting asset validation report
(new) basic setup of columns and data flow for asset validation
(new) write report on server to staging/tmp in CSV format

(fix) cleanup typescript error for checking if job is running
(fix) consolidated styles removing obsolete
(fix) moved shared types and styles to 'shared' folder
