+++ server/http/routes/api/generateDownloads.ts
@@ -95,7 +95,7 @@ const createOpForScene = async (idSystemObject: number, idUser: number): Promise
     LOG.info(`API.generateDownloads post creation. (result: ${H.Helpers.JSONStringify(result)})`,LOG.LS.eDEBUG);
 
     const isValid: boolean = result.data.isValid ?? false;
-    const isJobRunning: boolean = (result.data.activeJobs?.length>0) ?? false;
+    const isJobRunning: boolean = (!result.data.activeJobs) ? false : (result.data.activeJobs.length>0);
     const idWorkflow: number | undefined = (result.data.workflow?.idWorkflow) ?? undefined;
     const idWorkflowReport: number | undefined = (result.data.workflowReport?.idWorkflowReport) ?? undefined;
 
@@ -105,7 +105,6 @@ const createOpForScene = async (idSystemObject: number, idUser: number): Promise
         return generateResponse(false,result.message,idSystemObject,{ isValid, isJobRunning, idWorkflow, idWorkflowReport });
     }
 
-    console.log('a');
     return generateResponse(true,`Generating Downloads for: ${scene.Name}`,idSystemObject,{ isValid, isJobRunning, idWorkflow, idWorkflowReport });
 };
 const getOpStatusForScene = async (idSystemObject: number): Promise<GenDownloadsResponse> => {
